+++ src/ui/MoreDFSprite.cpp
@@ -7,7 +7,7 @@ bool MoreDFSprite::init() {
     return true;
 } 
 
-MoreDFSprite* MoreDFSprite::createWithStarCount(int starCount, bool smallSpr) {
+MoreDFSprite* MoreDFSprite::createWithStarCount(int starCount, bool smallSpr, int epic) {
     auto ret = new MoreDFSprite();
 
     std::string extraStuff = "";
@@ -16,9 +16,19 @@ MoreDFSprite* MoreDFSprite::createWithStarCount(int starCount, bool smallSpr) {
         extraStuff = "Small";
     }
 
-    if (Mod::get()->getSettingValue<bool>("legacy-difficulties")) {
+    if (Mod::get()->getSettingValue<bool>("legacy-difficulties") && !Mod::get()->getSettingValue<bool>("godlike-faces")) {
         std::string extraStuffPre = extraStuff;
         extraStuff = fmt::format("{}_Legacy", extraStuffPre);
+    }
+
+    if (!Mod::get()->getSettingValue<bool>("legacy-difficulties") && Mod::get()->getSettingValue<bool>("godlike-faces")) {
+        if (epic == 2) {
+            std::string extraStuffPre = extraStuff;
+            extraStuff = fmt::format("{}_Legendary", extraStuffPre);
+        } else if (epic == 3) {
+            std::string extraStuffPre = extraStuff;
+            extraStuff = fmt::format("{}_Mythic", extraStuffPre);
+        }
     }
 
     std::string spriteName = fmt::format("uproxide.more_difficulties/MD_Difficulty0{}{}.png", starCount, extraStuff);
@@ -30,6 +40,7 @@ MoreDFSprite* MoreDFSprite::createWithStarCount(int starCount, bool smallSpr) {
     if ((starCount == 4 && !SaveThings::casual) || (starCount == 7 && !SaveThings::tough) || (starCount == 9 && !SaveThings::cruel)) {
         return nullptr;
     }
+    
 
     if (ret && ret->initWithSpriteFrameName(spriteName.c_str())) {
         ret->autorelease();