+++ src/ui/MoreDFSprite.hpp
@@ -5,7 +5,7 @@ using namespace geode::prelude;
 class MoreDFSprite : public CCSprite {
     public:
         bool init();
-        static MoreDFSprite* createWithStarCount(int starCount, bool smallSpr);
+        static MoreDFSprite* createWithStarCount(int starCount, bool smallSpr, int epic);
         static MoreDFSprite* createWithName(std::string name, bool smallSpr);
         static MoreDFSprite* createWithLevelID(int levelID, std::string name, bool smallSpr);
         static MoreDFSprite* createWithSpriteFrameName(const char* name);